+++ features/overlays/seaCreaturesHpTracker.js
@@ -1,58 +1,86 @@
-import settings, { allOverlaysGui } from "../../settings";
-import { BOLD, YELLOW } from "../../constants/formatting";
-import { EntityArmorStand } from "../../constants/javaTypes";
+import settings, { allOverlaysGui, seaCreaturesHpOverlayGui } from "../../settings";
+import { AQUA, BOLD } from "../../constants/formatting";
 import { overlayCoordsData } from "../../data/overlayCoords";
 import { getWorldName, isInSkyblock } from "../../utils/playerState";
-import { BACKWATER_BAYOU, CRIMSON_ISLE, JERRY_WORKSHOP, WATER_HOTSPOT_WORLDS } from "../../constants/areas";
+import { BACKWATER_BAYOU, CRIMSON_ISLE, CRYSTAL_HOLLOWS, JERRY_WORKSHOP, WATER_FISHING_WORLDS, WATER_HOTSPOT_WORLDS } from "../../constants/areas";
 import { OFF_SOUND_MODE } from "../../constants/sounds";
 import { registerIf } from "../../utils/registers";
+import { getSeaCreaturesInRange } from "../../utils/entityDetection";
 
 const LOOTSHARE_DISTANCE = 30;
 const TRACKED_MOBS = [
     {
         worlds: [CRIMSON_ISLE],
-        mobShortName: 'Fiery Scuttler',
+        baseMobName: 'Fiery Scuttler',
     },
     {
         worlds: [CRIMSON_ISLE],
-        mobShortName: 'Lord Jawbus',
+        baseMobName: 'Lord Jawbus',
     },
     {
         worlds: [CRIMSON_ISLE],
-        mobShortName: 'Thunder',
+        baseMobName: 'Thunder',
     },
     {
         worlds: [CRIMSON_ISLE],
-        mobShortName: 'Plhlegblast',
+        baseMobName: 'Plhlegblast',
     },
     {
         worlds: [CRIMSON_ISLE],
-        mobShortName: 'Ragnarok',
+        baseMobName: 'Ragnarok',
     },
     {
         worlds: [JERRY_WORKSHOP],
-        mobShortName: 'Reindrake',
+        baseMobName: 'Reindrake',
     },
     {
         worlds: [JERRY_WORKSHOP],
-        mobShortName: 'Yeti',
+        baseMobName: 'Yeti',
     },
     {
         worlds: WATER_HOTSPOT_WORLDS,
-        mobShortName: 'Alligator',
+        baseMobName: 'Alligator',
     },
     {
         worlds: WATER_HOTSPOT_WORLDS,
-        mobShortName: 'Blue Ringed Octopus',
+        baseMobName: 'Blue Ringed Octopus',
     },
     {
         worlds: WATER_HOTSPOT_WORLDS,
-        mobShortName: 'Wiki Tiki',
+        baseMobName: 'Wiki Tiki',
+    },
+    {
+        worlds: WATER_HOTSPOT_WORLDS,
+        baseMobName: 'Wiki Tiki Laser Totem',
     },
     {
         worlds: [BACKWATER_BAYOU],
-        mobShortName: 'Titanoboa',
-    }
+        baseMobName: 'Titanoboa',
+    },
+    {
+        worlds: [CRYSTAL_HOLLOWS],
+        baseMobName: 'Abyssal Miner',
+    },
+    {
+        worlds: WATER_FISHING_WORLDS,
+        baseMobName: 'The Sea Emperor',
+    },
+    {
+        worlds: WATER_FISHING_WORLDS,
+        baseMobName: 'Water Hydra',
+    },
+    {
+        worlds: WATER_FISHING_WORLDS,
+        baseMobName: 'Phantom Fisher',
+    },
+    {
+        worlds: WATER_FISHING_WORLDS,
+        baseMobName: 'Grim Reaper',
+    },
+    {
+        worlds: WATER_FISHING_WORLDS,
+        baseMobName: 'Great White Shark',
+    },
 ];
 
 const TRACKED_WORLD_NAMES = TRACKED_MOBS
@@ -78,8 +106,6 @@ register("worldUnload", () => {
 
 function trackSeaCreaturesHp() {
     try {
-        const worldName = getWorldName();
-
         if (!settings.seaCreaturesHpOverlay ||
             !isInSkyblock() ||
             !TRACKED_WORLD_NAMES.includes(getWorldName())
@@ -87,25 +113,20 @@ function trackSeaCreaturesHp() {
             return;
         }
     
-        let currentMobs = [];
-        const entities = World.getAllEntitiesOfType(EntityArmorStand);
-    
-        const player = Player.getPlayer();
-        entities.forEach(entity => {
-            const name = entity?.getName();
-            const plainName = entity?.getName()?.removeFormatting();
-    
-            if (plainName.includes('[Lv') && plainName.includes('❤') && // Distinguish mobs from pets (e.g. Squid)
-                TRACKED_MOBS.filter(m => m.worlds.includes(worldName)).some(m => plainName.includes(m.mobShortName)) &&
-                (plainName.includes('Reindrake') || entity.distanceTo(player) <= LOOTSHARE_DISTANCE)
-            ) {
-                // Original nametag: §e﴾ §8[§7Lv400§8] §c§lThunder§r§r §e17M§f/§a35M§c❤ §e﴿ §b✯
-                const cleanName = name.replace('§e﴾ ', '').replace(' §e﴿', '').trim().split('] ')[1];
-                currentMobs.push(cleanName);          
-            }
-        });
-    
-        if (currentMobs.length > mobs.length && settings.soundMode !== OFF_SOUND_MODE) {
+        const currentMobs = getSeaCreaturesInRange(TRACKED_MOBS.map(n => n.baseMobName), LOOTSHARE_DISTANCE)
+            .sort((a, b) => a.currentHpNumber - b.currentHpNumber) // Lowest HP comes first
+            .map(sc => ({ nametag: sc.shortNametag, baseMobName: sc.baseMobName }))
+            .slice(0, settings.seaCreaturesHpOverlay_maxCount); // Top N
+
+        const addedMobNames = currentMobs.filter(cm => {
+            return !mobs.some(m => m.baseMobName === cm.baseMobName);
+         });
+
+        if (
+            currentMobs.length > mobs.length &&
+            settings.soundMode !== OFF_SOUND_MODE &&
+            !addedMobNames.every(m => m.baseMobName === 'Reindrake') // Reindrake flies around and goes out of nametags render distance periodically, we don't need sound for it
+        ) {
             World.playSound('random.orb', 0.75, 1);
         }
     
@@ -118,7 +139,6 @@ function trackSeaCreaturesHp() {
 
 function renderHpOverlay() {
     if (!settings.seaCreaturesHpOverlay ||
-        !mobs.length ||
         !isInSkyblock() ||
         !TRACKED_WORLD_NAMES.includes(getWorldName()) ||
         allOverlaysGui.isOpen()
@@ -126,13 +146,24 @@ function renderHpOverlay() {
         return;
     }
 
-    let overlayText = `${YELLOW}${BOLD}Sea creatures HP\n`;
-    mobs.forEach((mob) => {
-        overlayText += `${mob}\n`;
-    });
+    if (!mobs.length && seaCreaturesHpOverlayGui.isOpen()) {
+        const overlayText = `${AQUA}${BOLD}Sea creatures HP`;
+        drawText(overlayText);
+        return;
+    }
+    
+    if (mobs.length) {
+        let overlayText = `${AQUA}${BOLD}Sea creatures HP\n`;
+        mobs.forEach((mob) => {
+            overlayText += `${mob.nametag}\n`;
+        });
+        drawText(overlayText);
+    }
 
-    const overlay = new Text(overlayText, overlayCoordsData.seaCreaturesHpOverlay.x, overlayCoordsData.seaCreaturesHpOverlay.y)
-        .setShadow(true)
-        .setScale(overlayCoordsData.seaCreaturesHpOverlay.scale);
-    overlay.draw();
+    function drawText(overlayText) {
+        const overlay = new Text(overlayText, overlayCoordsData.seaCreaturesHpOverlay.x, overlayCoordsData.seaCreaturesHpOverlay.y)
+            .setShadow(true)
+            .setScale(overlayCoordsData.seaCreaturesHpOverlay.scale);
+        overlay.draw();
+    }
 }